Make skipping indexing of existing documents optional
Change I1fd78958f6fa0848f81630abab0428520f412baa introduced a feature
to skip indexing of changes which were already present in the index.
This prevented unnecessary work, if Gerrit had to be shutdown while
performing an online reindex. However, this change also broke offline
reindexing, since this forcefully deleted all documents in the index
before reindexing, which wasn't transparent to the indexers.

This change is a reimplementation of the feature allowing to skip existing
documents in the index. Skipping of indexing changes is now optional. For
online reindexing the `index.skipExistingDocuments` option has to
be set in the `gerrit.config`. For offline reindexing, the
`--skip-existing` flag can be used.

The issue with offline reindexing is fixed, by not emptying the index
before indexing, if `--skip-existing` has been set.

Test plan:
- offline reindexing:
  - with version upgrade
  - same version, already complete index -> successful
  - same version, incomplete index
  - same version, stale documents
- online reindexing
  - with version upgrade
  - same version, already complete index
  - same version, incomplete index
  - same version, stale documents

Todos:
- Automated tests

+++ java/com/google/gerrit/pgm/Reindex.java
@@ -98,6 +98,9 @@ public class Reindex extends SiteProgram {
   @Option(name = "--build-bloom-filter", usage = "Build bloom filter for H2 disk caches.")
   private boolean buildBloomFilter;
 
+  @Option(name = "--skip-existing", usage = "Skips documents already existing in index.")
+  private boolean skipExisting;
+
   private Injector dbInjector;
   private Injector sysInjector;
   private Injector cfgInjector;
@@ -258,12 +261,14 @@ public class Reindex extends SiteProgram {
     requireNonNull(
         index, () -> String.format("no active search index configured for %s", def.getName()));
     index.markReady(false);
-    index.deleteAll();
+    if (!skipExisting) {
+      index.deleteAll();
+    }
 
     SiteIndexer<K, V, I> siteIndexer = def.getSiteIndexer();
     siteIndexer.setProgressOut(System.err);
     siteIndexer.setVerboseOut(verbose ? System.out : NullOutputStream.INSTANCE);
-    SiteIndexer.Result result = siteIndexer.indexAll(index);
+    SiteIndexer.Result result = siteIndexer.indexAll(index, skipExisting);
     int n = result.doneCount() + result.failedCount();
     double t = result.elapsed(TimeUnit.MILLISECONDS) / 1000d;
     System.out.format(
